app/requirements.prod.txt
Django == 4.2.1
djangorestframework == 3.14
pywebpush == 1.14.0
django-cors-headers == 3.13.0
aiosmtplib == 2.0.1
django-environ == 0.10.0
adrf == 0.1.1
uvicorn == 0.22.0
gunicorn == 20.1.0
